refactor: move ValidatorsComposite to its own file
