+++ src/application/validators/composite.ts
@@ -1,0 +1,15 @@
+import { Validator } from './validator'
+
+export class ValidatorsComposite implements Validator {
+  constructor (private readonly validators: Validator[]) {}
+
+  validate (): Error | undefined {
+    for (const validator of this.validators) {
+      const error = validator.validate()
+
+      if (error !== undefined) {
+        return error
+      }
+    }
+  }
+}
