+++ src/application/validators/index.ts
@@ -1,2 +1,3 @@
 export * from './required-field'
 export * from './validator'
+export * from './composite'
